// src/app/shared/enum/permission.enum.ts
export const enum PermissionEnum {
  crmLogin = 'crmLogin',
  accountRead = 'accountRead',
  accountWrite = 'accountWrite',
  accountAdmin = 'accountAdmin',
  accountMyProfile = 'accountMyProfile',
  userSwitch = 'userSwitch',
  dashboardRead = 'dashboardRead',
  propertyRead = 'propertyRead',
  propertyReadSale = 'propertyReadSale',
  propertyReadRental = 'propertyReadRental',
  propertyWrite = 'propertyWrite',
  propertyDelete = 'propertyDelete',
  contactRead = 'contactRead',
  contactWrite = 'contactWrite',
  contactDelete = 'contactDelete',
  contactExport = 'contactExport',
  contactTransferActivity = 'contactTransferActivity',
  leadRead = 'leadRead',
  leadWrite = 'leadWrite',
  leadManager = 'leadManager',
  taskRead = 'taskRead',
  taskWrite = 'taskWrite',
  taskManager = 'taskManager',
  taskDelete = 'taskDelete',
  matchingRead = 'matchingRead',
  promotionRead = 'promotionRead',
  promotionWrite = 'promotionWrite',
  promotionDelete = 'promotionDelete',
  agendaRead = 'agendaRead',
  mailboxRead = 'mailboxRead',
  mailboxWrite = 'mailboxWrite',
  mailboxTemplateRead = 'mailboxTemplateRead',
  mailboxTemplateWrite = 'mailboxTemplateWrite',
  mailingSendOnBehalf = 'mailingSendOnBehalf',
  reportingRead = 'reportingRead',
  reportingManager = 'reportingManager',
  mlsRead = 'mlsRead',
  mlsWrite = 'mlsWrite',
  websiteConfigRead = 'websiteConfigRead',
  websiteConfigWrite = 'websiteConfigWrite',
  websiteContentRead = 'websiteContentRead',
  websiteContentWrite = 'websiteContentWrite',
  agencyProfileRead = 'agencyProfileRead',
  agencyProfileWrite = 'agencyProfileWrite',
  agencyGatewayRead = 'agencyGatewayRead',
  agencyGatewayWrite = 'agencyGatewayWrite',
  agencyMarketingRead = 'agencyMarketingRead',
  agencyMarketingWrite = 'agencyMarketingWrite',
  agencyProductionRead = 'agencyProductionRead',
  agencyProductionWrite = 'agencyProductionWrite',
  agencyDeviceRead = 'agencyDeviceRead',
  agencyBackupRead = 'agencyBackupRead',
  propertyPublishWebsite = 'propertyPublishWebsite',
  propertyPublishPortal = 'propertyPublishPortal',
  propertyMortgageRead = 'propertyMortgageRead',
  propertyMortgageWrite = 'propertyMortgageWrite',
  agencyPreferenceRead = 'agencyPreferenceRead',
  agencyPreferenceWrite = 'agencyPreferenceWrite',
  sectorRead = 'sectorRead',
  sectorWrite = 'sectorWrite',
  sectorDelete = 'sectorDelete',
  historyRead = 'historyRead',
  customAttributeRead = 'customAttributeRead',
  customAttributeWrite = 'customAttributeWrite',
  domainNameRead = 'domainNameRead',
  contractRead = 'contractRead',
  contractWrite = 'contractWrite',
  websiteRead = 'websiteRead',
  agencyGroupAdmin = 'agencyGroupAdmin',
  taskCalendarExport = 'taskCalendarExport',
  matchingMarginRead = 'matchingMarginRead',
  matchingMarginWrite = 'matchingMarginWrite',
  restrictionRead = 'restrictionRead',
  restrictionWrite = 'restrictionWrite',
  restrictionAdmin = 'restrictionAdmin',
  suggestionRead = 'suggestionRead',
  suggestionWrite = 'suggestionWrite',
}