// src/app/shared/enum/operation.enum.ts
export const enum OperationEnum {
  propertyAddBasket = 'property-add-basket',
  propertyRemoveBasket = 'property-remove-basket',
  propertyGenerateReport = 'property-generate-report',
  propertySendEmail = 'property-send-email',
  propertyDuplicate = 'property-duplicate',
  propertyConvertTypeRent = 'property-convert-type-rent',
  propertyConvertTypeSell = 'property-convert-type-sell',
  propertyRemoveMls = 'property-remove-mls',
  propertyManagePublication = 'property-manage-publication',
  propertyTransferBroker = 'property-transfer-broker',
  propertyArchive = 'property-archive',
  propertyUnarchive = 'property-unarchive',
  reportingSend = 'reporting-send',
  reportingRefuse = 'reporting-refuse',
  promotionRemoveMls = 'promotion-remove-mls',
  promotionSendEmail = 'promotion-send-email',
  promotionArchive = 'promotion-archive',
  promotionUnarchive = 'promotion-unarchive',
  leadSendEmail = 'lead-send-email',
  leadModifyStatus = 'lead-modify-status',
  contactAddBasket = 'contact-add-basket',
  contactRemoveBasket = 'contact-remove-basket',
  contactSendEmail = 'contact-send-email',
  contactModifyBroker = 'contact-modify-broker',
  contactTransferActivity = 'contact-transfer-activity',
  contactTransferBroker = 'contact-transfer-broker',
  contactArchive = 'contact-archive',
  contactUnarchive = 'contact-unarchive',
  contactExportFull = 'contact-export-full',
  contactExportSummary = 'contact-export-summary',
  matchingWaiting = 'matching-waiting',
  sectorArchive = 'sector-archive',
  sectorUnarchive = 'sector-unarchive',
  reportSendEmail = 'report-send-email',
  customAttributeDelete = 'customAttributeDelete',
  userProfile = 'user-profile',
  userSwitch = 'user-switch',
  userHelp = 'user-help',
  userSignout = 'user-signout',
  taskSendEmail = 'task-send-email',
  taskAddBasket = 'task-add-basket',
  taskRemove = 'task-remove',
  taskMarkFinish = 'task-mark-finish',
  suggestionList = 'suggestion-list',
}